django_project/dog_model_service/requirements.txt
bs4
djangorestframework
lxml
mysqlclient
pandas
pip-chill
requests
sklearn
opencv-python
Pillow
tensorflow